+++ src/lib.rs
@@ -105,6 +105,7 @@
     clippy::inline_always
 )]
 
+pub mod constraint_graph;
 pub mod device;
 #[cfg(feature = "fix128-arithmetic")]
 pub mod fix128;